Service for User and Role

+++ src/main/java/com/ideas2it/ideas2movie/model/Role.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Ideas2IT Technologies. All rights reserved.
+ * Copyright 2023 Ideas2IT Technologies. All rights reserved.
  * IDEAS2IT PROPRIETARY/CONFIDENTIAL.
  */
 package com.ideas2it.ideas2movie.model;
@@ -36,7 +36,7 @@ import lombok.Setter;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @NotNull
     private String name;
 }
